requirements.txt
pandas
tqdm
MLB-StatsAPI
flumine
chardet
ipython
ipykernel